Prototype pulling data from mapbox vector tiles
The support in unity/mono for this isn't quite ready. Mapbox vector tiles are sent using protobuffers, which have a C# implementation, but the instructions point to installing that via nuget, which isn't supported in monodevelop yet. So, for now it's probably best to stick with reading in the XML file.